fix(tools/gasbench): address PR review findings (NaN bounds, lint gate)
- nullableBound treats NaN as non-finite alongside ±Inf (encoding/json rejects
  both), and Underpowered detection uses the same finite() predicate — closes
  the D-1 gap Cursor flagged. benchmath's median CI emits ±Inf sentinels, never
  NaN, so this was latent, but the wire invariant is "non-finite never reaches
  the encoder," now enforced and unit-pinned (distinguishable(NaN,NaN)=false,
  nullableBound(NaN)=nil).
- Drop the redundant int64() conversions in rusageSnapshot: Rusage.Nvcsw/Nivcsw
  are int64 on every target platform (darwin/arm64, linux/amd64, linux/arm64) —
  clears unconvert.
- Guard BuildCaseWith's conversions for gosec G115: reps > 0 and Arity <= 16
  (DUP16 is the deepest lift) now panic at construction, with nolint on the
  guarded sites. golangci-lint run ./tools/gasbench/...: 0 issues.

+++ tools/gasbench/gasbench.go
@@ -70,7 +70,7 @@ func rusageSnapshot() (nvcsw, nivcsw int64, err error) {
 	if err := syscall.Getrusage(syscall.RUSAGE_SELF, &ru); err != nil {
 		return 0, 0, fmt.Errorf("gasbench: getrusage: %w", err)
 	}
-	return int64(ru.Nvcsw), int64(ru.Nivcsw), nil
+	return ru.Nvcsw, ru.Nivcsw, nil
 }
 
 // Measure runs Warmup discarded iterations, then Iterations timed iterations
